+++ src/Interpreter/NLInteger.java
@@ -13,7 +13,7 @@ class NLInteger extends Value {
         if (value instanceof NLString)
             return new NLString(this.value + value.getValue());
         else if (value instanceof NLFloat)
-            return new NLFloat((float) this.value + Float.parseFloat(value.getValue()));
+            return new NLFloat((double) this.value + Double.parseDouble(value.getValue()));
         else
             return new NLInteger(this.value + Integer.parseInt(value.getValue()));
     }
